+++ lib/QECCore/src/interface.jl
@@ -117,3 +117,42 @@ Abstract type representing algorithms for computing
 the minimum distance of quantum error correction codes.
 """
 abstract type AbstractDistanceAlg end
+
+"""
+    metacheck_matrix_x(c::AbstractCSSCode)
+
+Returns the `X`-metacheck matrix (``\\partial_{Mx}`` boundary map in
+[chain complex](https://en.wikipedia.org/wiki/Chain_complex) notation) for a CSS code.
+
+This matrix verifies validity of `X`-syndromes (`Z`-error measurements).
+
+Only CSS codes built using chain complexes and homology have this method.
+
+See also: [`metacheck_matrix_z`](@ref), [`metacheck_matrix`](@ref), [`parity_matrix_x`](@ref)
+"""
+function metacheck_matrix_x end
+
+"""
+    metacheck_matrix_z(c::AbstractCSSCode)
+
+Returns the `Z`-metacheck matrix (``\\partial_{Mz}`` boundary map in
+[chain complex](https://en.wikipedia.org/wiki/Chain_complex) notation) for a CSS code.
+
+This matrix verifies validity of `Z`-syndromes (`X`-error measurements).
+
+Only CSS codes built using chain complexes and homology have this method.
+
+See also: [`metacheck_matrix_x`](@ref), [`metacheck_matrix`](@ref), [`parity_matrix_z`](@ref)
+"""
+function metacheck_matrix_z end
+
+"""
+    metacheck_matrix(c::AbstractCSSCode)
+
+Returns both `X` and `Z` metacheck matrices.
+
+Only CSS codes built using chain complexes and homology have this method.
+
+See also: [`metacheck_matrix_x`](@ref), [`metacheck_matrix_z`](@ref)
+"""
+function metacheck_matrix end