+++ lib/QECCore/CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## v0.1.2 - dev
 
+- introduce `metacheck_matrix_x`, `metacheck_matrix_z`, and `metacheck_matrix` for CSS codes built using chain complexes and homology.
 - Add `ReedMuller`, `RecursiveReedMuller`, and `QuantumReedMuller` from `QuantumClifford.ECC`.
 
-  
 ## v0.1.1
 
 - Add `AbstractDistanceAlg` for dispatch for `distance`